[PATCH] Fix many NPEs and IOOBEs
It's staggering how many exceptions there were, but in the end my application that I used to test this against was decompiled without any significant errors (there's still the issue of <unknown> being used in enchanted for-loops) and correct logic (compared to the output of Krakatau). Of course I'd now need to figure out how to decompile enums.

Co-Authored-By: Geolykt <admin@geolykt.de>