Save method bytecode for later parsing
Instead of discarding the bytecode and exception handler data this saves it in the StructCodeAttribute so it doesn't have to get looked up from the class again.

This prevents a crash where the wrong bytecode is used in case of duplicate methods.

Only one version of the duplicate method will show up in the output (twice), but at least the rest of the class is decompiled correctly.