Fix variable liveness iteration algorithm implemented improperly
- all live maps now copy the sforms properly, with a deep copy instead of a shallow copy. This removes the memory dependency between the live sets, making them properly independent.
- when live variables are iterated when stack variables are replaced, the current node's live map is properly modified. Previously it used the iterator but that didn't actually do any removing. Seems like an old algorithm that was never changed?
Effects: This completely eliminates liveverset being empty, so this will have to be reverted to see what it's doing with the unknown variables.
When running it over minecraft, seems like it allows IfHelper to restructure else ifs to be in a better orientation now? I have no clue at all how that happened but it makes the decompile better, so it's nice