Fix field array ppmm, create framework for multi-pass stack var simplification to handle certain bytecode cases better
Creates a second pass for stack simplification used currently to more aggressively inline across duplicated bytecode. The second pass only happens when the first pass wasn't able to inline anything, making sure it doesn't break existing code gen. This is configurable with a fernflower option as it may do strange things with obfuscated bytecode, though the chance for that is low. The algorithm is inspired by DomHelper's use of 2 passes to calculate postdominators differently, to try find find subgraphs if the first pass failed.