PPMM inlining to help stack var simplification
Not perfect, but initial attempt. Fixes cases where stack vars aren't properly inlined due to ppi and mmi getting in the way. Still more work to be done though! The algorithm used here is notably imperfect and hardcoded to not handle a bunch of edge cases, so a better one would be a good thing to add in the future.