Be more forgiving when parsing arguments for paths
Instead of counting the non-options and using that to determine the last
file path for the destination, instead store the last file path and
process that lazily.

This accounts for the cases where a single 'pseudo-options' (--only,
--silent, --add-external) is specified alongside a single file path,
where it previously incorrectly determined the file path to be the
destination.

This also accounts for the case where pseudo-options are specified at
the end of the command-line invocation, rather than before all file
paths.

Fixes #492