test: Test `lcmp`, `fcmp` and `dcmp` to `boolean` in TestNumberCompareToBoolean
All the tests are functionally `number == number && number != 0`, baking in a NaN check.

For the extra curious, the NaN checks in the floating point tests work by taking advantage
of the same fcmp insn will always send NaN to 1 or -1, allowing for xor to collapse it to 0.

`iconst_1`, `iand` are evidently redundant but can be helpful if you need to strictly fit in one bit.